Code/requirements.txt
torch
transformers
peft
pillow
matplotlib